Client

This page contains documentation of the public API of the Julia client. In the Julia REPL one can access this documentation by entering the help mode with ? and then writing the function for which the documentation should be shown.

RedPitaya

RedPitayaDAQServer.RedPitaya — Type

RedPitaya(ip [, port = 5025, isMaster = false])

Create a RedPitaya (i.e. a connection to a RedPitayaDAQServer server) which is the central structure of the Julia client library.

Used to communicate with the server and manage the connection and metadata. During the construction the connection is established and the calibration values are loaded from the RedPitayas EEPROM. Throws an error if a timeout occurs while attempting to connect.

Examples

julia> rp = RedPitaya("192.168.1.100");

julia> decimation(rp, 8)

julia> decimation(rp)
8

RedPitayaDAQServer.query — Function

query(rp::RedPitaya, cmd [, timeout = 5.0, N = 100])

Send a query to the RedPitaya command socket. Return reply as String.

Waits for timeout seconds and checks every timeout/N seconds.

See also receive.

query(rp::RedPitaya, cmd, T::Type [timeout = 5.0, N = 100])

Send a query to the RedPitaya. Parse reply as T.

Waits for timeout seconds and checks every timeout/N seconds.

RedPitayaDAQServer.receive — Function

receive(rp)

Receive a String from the RedPitaya command socket. Reads until a whole line is received

ADC

RedPitayaDAQServer.decimation — Function

decimation(rp::RedPitaya)

Return the decimation of the RedPitaya.

Examples

julia> rp = RedPitaya("192.168.1.100");

julia> decimation!(rp, 8)

julia> decimation(rp)
8

  decimation(rpc::RedPitayaCluster, value)

As single RedPitaya, but applied to only the master. See decimation

RedPitayaDAQServer.numChan — Function

numChan(rp::RedPitaya)

Return the number of ADC channel of a RedPitaya.

RedPitayaDAQServer.samplesPerPeriod — Function

samplesPerPeriod(rp::RedPitaya)

Return the number of samples per period.

Example

julia> samplesPerPeriod!(rp, 256)

julia> samplesPerPeriod(rp)
256

  samplesPerPeriod(rpc::RedPitayaCluster, value)

As single RedPitaya, but applied to only the master. See samplesPerPeriod

RedPitayaDAQServer.periodsPerFrame — Function

periodsPerFrame(rp::RedPitaya)

Return the number of periods per frame.

Example

julia> periodsPerFrame!(rp, 16)

julia> periodsPerFrame(rp)
16

  periodsPerFrame(rpc::RedPitayaCluster, value)

As single RedPitaya, but applied to only the master. See periodsPerFrame

RedPitayaDAQServer.currentWP — Function

currentWP(rp::RedPitaya)

Return the current writepointer of the RedPitaya.

RedPitayaDAQServer.currentFrame — Function

currentFrame(rp::RedPitaya)

Return the current frame of the RedPitaya based on the current writepointer, samples per period and periods per frame.

See also currentWP, samplesPerPeriod, periodsPerFrame.

RedPitayaDAQServer.currentPeriod — Function

currentPeriod(rp::RedPitaya)

Return the current period of the RedPitaya based on the current writepointer and samples per period.

See also currentWP, samplesPerPeriod.

DAC

Cluster