Client

This page contains documentation of the public API of the Julia client. In the Julia REPL one can access this documentation by entering the help mode with ? and then writing the function for which the documentation should be shown.

RedPitaya

RedPitayaDAQServer.RedPitaya — Type

RedPitaya

Struct representing a connection to a RedPitayaDAQServer.

Contains the sockets used for communication and connection related metadata. Also contains fields for client specific concepts such as periods, frames and calibration values.

Sockets.connect — Function

connect(rpc::RedPitayaCluster)

As with single RedPitaya, but applied to all RedPitayas in a cluster.

RedPitayaDAQServer.disconnect — Function

disconnect(rpc::RedPitayaCluster)

As with single RedPitaya, but applied to all RedPitayas in a cluster.

Sockets.send — Function

send(rp, cmd)

Send a command to the RedPitaya

RedPitayaDAQServer.query — Function

query(rp::RedPitaya, cmd [, timeout = 5.0, N = 100])

Send a query to the RedPitaya command socket. Return reply as String.

Waits for timeout seconds and checks every timeout/N seconds.

See also receive.

query(rp::RedPitaya, cmd, T::Type [timeout = 5.0, N = 100])

Send a query to the RedPitaya. Parse reply as T.

Waits for timeout seconds and checks every timeout/N seconds.

RedPitayaDAQServer.receive — Function

receive(rp)

Receive a String from the RedPitaya command socket. Reads until a whole line is received

RedPitayaDAQServer.RedPitayaCluster — Type

RedPitayaCluster

Struct representing a cluster of RedPitayas. Such a cluster should share a common clock and master trigger.

The structure implements the indexing and iterable interfaces.

RedPitayaDAQServer.serverMode — Function

serverMode(rpc::RedPitayaCluster, value)

As with single RedPitaya, but applied to only the master.

RedPitayaDAQServer.serverMode! — Function

  serverMode!(rpc::RedPitayaCluster, value)

As with single RedPitaya, but applied to all RedPitayas in a cluster.

ADC Configuration

Missing docstring.

Missing docstring for RedPitayaDAQServer.TriggerMode. Check Documenter's build log for details.

RedPitayaDAQServer.triggerMode — Function

triggerMode(rpc::RedPitayaCluster, value)

As with single RedPitaya, but applied to only the master.

RedPitayaDAQServer.triggerMode! — Function

triggerMode(rp::RedPitaya, mode::String)

Set the trigger mode of the RedPitaya. Valid values are "INTERNAL" or "EXTERNAL".

triggerMode(rp::RedPitaya, mode::String)

Set the trigger mode of the RedPitaya.

  triggerMode!(rpc::RedPitayaCluster, value)

As with single RedPitaya, but applied to all RedPitayas in a cluster.

RedPitayaDAQServer.keepAliveReset — Function

keepAliveReset(rp::RedPitaya)

Determine whether the keepAliveReset is set.

keepAliveReset(rpc::RedPitayaCluster, value)

As with single RedPitaya, but applied to only the master.

RedPitayaDAQServer.keepAliveReset! — Function

keepAliveReset!(rp::RedPitaya, val::Bool)

Set the keepAliveReset to val.

  keepAliveReset!(rpc::RedPitayaCluster, value)

As with single RedPitaya, but applied to all RedPitayas in a cluster.

RedPitayaDAQServer.decimation — Function

decimation(rp::RedPitaya)

Return the decimation of the RedPitaya.

Examples

julia> rp = RedPitaya("192.168.1.100");

julia> decimation!(rp, 8)

julia> decimation(rp)
8

decimation(rpc::RedPitayaCluster, value)

As with single RedPitaya, but applied to only the master.

RedPitayaDAQServer.decimation! — Function

decimation!(rp::RedPitaya, dec)

Set the decimation of the RedPitaya.

Examples

julia> decimation!(rp, 8)

julia> decimation(rp)
8

  decimation!(rpc::RedPitayaCluster, value)

As with single RedPitaya, but applied to all RedPitayas in a cluster.

RedPitayaDAQServer.numChan — Function

numChan(rp::RedPitaya)

Return the number of ADC channel of a RedPitaya.

RedPitayaDAQServer.samplesPerPeriod — Function

samplesPerPeriod(rp::RedPitaya)

Return the number of samples per period.

Example

julia> samplesPerPeriod!(rp, 256)

julia> samplesPerPeriod(rp)
256

samplesPerPeriod(rpc::RedPitayaCluster, value)

As with single RedPitaya, but applied to only the master.

RedPitayaDAQServer.samplesPerPeriod! — Function

samplesPerPeriod!(rp::RedPitaya, value)

Set the number of samples per period.

Example

julia> samplesPerPeriod!(rp, 256)

julia> samplesPerPeriod(rp)
256

  samplesPerPeriod!(rpc::RedPitayaCluster, value)

As with single RedPitaya, but applied to all RedPitayas in a cluster.

RedPitayaDAQServer.periodsPerFrame — Function

periodsPerFrame(rp::RedPitaya)

Return the number of periods per frame.

Example

julia> periodsPerFrame!(rp, 16)

julia> periodsPerFrame(rp)
16

periodsPerFrame(rpc::RedPitayaCluster, value)

As with single RedPitaya, but applied to only the master.

RedPitayaDAQServer.periodsPerFrame! — Function

periodsPerFrame(rp::RedPitaya, value)

Set the number of periods per frame.

Example

julia> periodsPerFrame!(rp, 16)

julia> periodsPerFrame(rp)
16

  periodsPerFrame!(rpc::RedPitayaCluster, value)

As with single RedPitaya, but applied to all RedPitayas in a cluster.

DAC Configuration

RedPitayaDAQServer.amplitudeDAC — Function

amplitudeDAC(rp::RedPitaya, channel, component)

Return the amplitude of composite waveform component for channel.

See amplitudeDAC!.

Examples

julia> amplitudeDAC!(rp, 1, 1, 0.5);

julia> amplitudeDAC(rp, 1, 1)
0.5

amplitudeDAC(rpc::RedPitayaCluster, chan::Integer, component::Integer)

As with single RedPitaya. The chan index refers to the total channel available in a cluster, two per RedPitaya. For example channel 4 would refer to the second channel of the second RedPitaya.

RedPitayaDAQServer.amplitudeDAC! — Function

amplitudeDAC!(rp::RedPitaya, channel, component, value)

Set the amplitude of composite waveform component for channel.

See amplitudeDAC.

Examples

julia> amplitudeDAC!(rp, 1, 1, 0.5);

julia> amplitudeDAC(rp, 1, 1)
0.5

amplitudeDAC!(rpc::RedPitayaCluster, chan::Integer, component::Integer, value)

As with single RedPitaya. The chan index refers to the total channel available in a cluster, two per RedPitaya. For example channel 4 would refer to the second channel of the second RedPitaya.

RedPitayaDAQServer.offsetDAC — Function

offsetDAC(rp::RedPitaya, channel)

Return the offset for channel.

See offsetDAC!.

Examples

julia> offsetDAC!(rp, 1, 0.2);

julia> offsetDAC(rp, 1)
0.2

offsetDAC(rpc::RedPitayaCluster, chan::Integer)

As with single RedPitaya. The chan index refers to the total channel available in a cluster, two per RedPitaya. For example channel 4 would refer to the second channel of the second RedPitaya.

RedPitayaDAQServer.offsetDAC! — Function

offsetDAC!(rp::RedPitaya, channel, value)

Set the offset for channel.

See offsetDAC.

Examples

julia> offsetDAC!(rp, 1, 0.2);

julia> offsetDAC(rp, 1)
0.2

offsetDAC!(rpc::RedPitayaCluster, chan::Integer, value)

As with single RedPitaya. The chan index refers to the total channel available in a cluster, two per RedPitaya. For example channel 4 would refer to the second channel of the second RedPitaya.

RedPitayaDAQServer.frequencyDAC — Function

frequencyDAC(rp::RedPitaya, channel, component)

Return the frequency of composite waveform component for channel.

See frequencyDAC!.

Examples

julia> frequencyDAC!(rp, 1, 1, 2400);

julia> frequencyDAC(rp, 1, 1)
2400

frequencyDAC(rpc::RedPitayaCluster, chan::Integer, component::Integer)

As with single RedPitaya. The chan index refers to the total channel available in a cluster, two per RedPitaya. For example channel 4 would refer to the second channel of the second RedPitaya.

RedPitayaDAQServer.frequencyDAC! — Function

frequencyDAC!(rp::RedPitaya, channel, component, value)

Set the frequency of composite waveform component for channel.

See frequencyDAC.

Examples

julia> frequencyDAC!(rp, 1, 1, 2400);

julia> frequencyDAC(rp, 1, 1)
2400

frequencyDAC!(rpc::RedPitayaCluster, chan::Integer, component::Integer, value)

As with single RedPitaya. The chan index refers to the total channel available in a cluster, two per RedPitaya. For example channel 4 would refer to the second channel of the second RedPitaya.

RedPitayaDAQServer.phaseDAC — Function

phaseDAC(rp::RedPitaya, channel, component)

Return the phase of composite waveform component for channel.

See phaseDAC!.

Examples

julia> phaseDAC!(rp, 1, 1, 0.0);

julia> phaseDAC(rp, 1, 0.0)
0.0

phaseDAC(rpc::RedPitayaCluster, chan::Integer, component::Integer)

As with single RedPitaya. The chan index refers to the total channel available in a cluster, two per RedPitaya. For example channel 4 would refer to the second channel of the second RedPitaya.

RedPitayaDAQServer.phaseDAC! — Function

phaseDAC!(rp::RedPitaya, channel, component, value)

Set the phase of composite waveform component for channel.

See phaseDAC.

Examples

julia> phaseDAC!(rp, 1, 1, 0.0);

julia> phaseDAC(rp, 1, 0.0)
0.0

phaseDAC!(rpc::RedPitayaCluster, chan::Integer, component::Integer, value)

As with single RedPitaya. The chan index refers to the total channel available in a cluster, two per RedPitaya. For example channel 4 would refer to the second channel of the second RedPitaya.

RedPitayaDAQServer.jumpSharpnessDAC — Function

jumpSharpnessDAC(rp::RedPitaya, channel, value)

Return the jumpSharpness of composite waveform for channel.

See jumpSharpnessDAC!.

Examples

julia> jumpSharpnessDAC!(rp, 1, 0.01);

julia> jumpSharpnessDAC(rp, 1)
0.01

jumpSharpnessDAC(rpc::RedPitayaCluster, chan::Integer)

As with single RedPitaya. The chan index refers to the total channel available in a cluster, two per RedPitaya. For example channel 4 would refer to the second channel of the second RedPitaya.

RedPitayaDAQServer.jumpSharpnessDAC! — Function

jumpSharpnessDAC!(rp::RedPitaya, channel, value)

Set the jumpSharpness of composite waveform for channel.

See jumpSharpnessDAC.

Examples

julia> jumpSharpnessDAC!(rp, 1, 0.01);

julia> jumpSharpnessDAC(rp, 1)
0.01

jumpSharpnessDAC!(rpc::RedPitayaCluster, chan::Integer, value)

As with single RedPitaya. The chan index refers to the total channel available in a cluster, two per RedPitaya. For example channel 4 would refer to the second channel of the second RedPitaya.

Missing docstring.

Missing docstring for RedPitayaDAQServer.SignalType. Check Documenter's build log for details.

RedPitayaDAQServer.signalTypeDAC — Function

signalTypeDAC!(rp::RedPitaya, channel, value)

Return the signalType of composite waveform for channel.

See signalTypeDAC!.

Examples

julia> signalTypeDAC!(rp, 1, SINE);

julia> signalTypeDAC(rp, 1)
SINE

signalTypeDAC(rpc::RedPitayaCluster, chan::Integer)

As with single RedPitaya. The chan index refers to the total channel available in a cluster, two per RedPitaya. For example channel 4 would refer to the second channel of the second RedPitaya.

RedPitayaDAQServer.signalTypeDAC! — Function

signalTypeDAC!(rp::RedPitaya, channel, value)

Set the signalType of composite waveform for channel.

See signalTypeDAC.

Examples

julia> signalTypeDAC!(rp, 1, SINE);

julia> signalTypeDAC(rp, 1)
SINE

signalTypeDAC!(rpc::RedPitayaCluster, chan::Integer, value)

As with single RedPitaya. The chan index refers to the total channel available in a cluster, two per RedPitaya. For example channel 4 would refer to the second channel of the second RedPitaya.

RedPitayaDAQServer.numSeqChan — Function

numSeqChan(rp::RedPitaya)

Return the number of sequence channel.

RedPitayaDAQServer.numSeqChan! — Function

numSeqChan(rp::RedPitaya, value)

Set the number of sequence channel. Valid values are between 1 and 4.

RedPitayaDAQServer.samplesPerStep — Function

samplesPerStep(rp::RedPitaya)

Return the number of samples per sequence step.

RedPitayaDAQServer.samplesPerStep! — Function

samplesPerStep!(rp::RedPitaya, value::Integer)

Set the number of samples per sequence step.

RedPitayaDAQServer.stepsPerFrame! — Function

stepsPerFrame!(rp::RedPitaya, stepsPerFrame)

Set the number of samples per steps s.t. stepsPerFrame sequence steps in a frame.

See samplesPerPeriod!, periodsPerFrame!, samplesPerStep!.

RedPitayaDAQServer.AbstractSequence — Type

AbstractSequence

Abstract struct of client-side representation of a sequence.

See appendSequence!, prepareSequence!, ArbitrarySequence.

RedPitayaDAQServer.appendSequence! — Function

appendSequence!(rp::RedPitaya, seq::AbstractSequence)

Transmit the client-side representation seq to the server and append it to the current list of sequences.

See prepareSequence!, clearSequences!.

  appendSequence!(rpc::RedPitayaCluster, value)

As with single RedPitaya, but applied to all RedPitayas in a cluster.

RedPitayaDAQServer.prepareSequence! — Function

prepareSequence!(rp::RedPitaya)

Instruct the server to prepare the currently added sequences. Returns true if server could prepare, false otherwise.

prepareSequence!(rpc::RedPitayaCluster)

As with single RedPitaya, but applied to all RedPitayas in a cluster. Returns true if all could servers could prepare, false otherwise.

RedPitayaDAQServer.clearSequences! — Function

clearSequences!(rp::RedPitaya)

Instruct the server to remove all sequences from its list.

RedPitayaDAQServer.popSequence! — Function

popSequence!(rp::RedPitaya)

Instruct the server to remove the last added sequence from its list.

Base.length — Function

length(seq::AbstractSequence)

Return the number of steps a sequence will take.

length(rpc::RedPitayaCluster)

Return the number of RedPitayas in cluster rpc.

RedPitayaDAQServer.start — Function

start(seq::AbstractSequence)

Return the number of steps after which a sequence leaves the ramp up phase.

RedPitayaDAQServer.ArbitrarySequence — Type

ArbitrarySequence <: AbstractSequence

Struct representing a sequence in which the server directly takes the values from the given LUT.

Measurement and Transmission

RedPitayaDAQServer.masterTrigger — Function

masterTrigger(rp::RedPitaya)

Determine whether the master trigger is set.

Example

julia> masterTrigger!(rp, true)

julia>masterTrigger(rp)
true

masterTrigger(rpc::RedPitayaCluster, value)

As with single RedPitaya, but applied to only the master.

RedPitayaDAQServer.masterTrigger! — Function

masterTrigger!(rp::RedPitaya, val::Bool)

Set the master trigger of the RedPitaya to val.

Example

julia> masterTrigger!(rp, true)

julia>masterTrigger(rp)
true

masterTrigger(rpc::RedPitayaCluster, val::Bool)

Set the master trigger of the cluster to val.

For val equals to true this is the same as calling the function on the RedPitaya returned by master(rpc). If val is false then the keepAliveReset is set to true for all RedPitayas in the cluster before the master trigger is disabled. Afterwards the keepAliveReset is set to false again.

See also master, keepAliveReset!.

RedPitayaDAQServer.currentWP — Function

currentWP(rp::RedPitaya)

Return the current writepointer of the RedPitaya.

currentWP(rpc::RedPitayaCluster, value)

As with single RedPitaya, but applied to only the master.

RedPitayaDAQServer.currentFrame — Function

currentFrame(rp::RedPitaya)

Return the current frame of the RedPitaya based on the current writepointer, samples per period and periods per frame.

See also currentWP, samplesPerPeriod, periodsPerFrame.

currentFrame(rpc::RedPitayaCluster, value)

As with single RedPitaya, but applied to only the master.

RedPitayaDAQServer.currentPeriod — Function

currentPeriod(rp::RedPitaya)

Return the current period of the RedPitaya based on the current writepointer and samples per period.

See also currentWP, samplesPerPeriod.

currentPeriod(rpc::RedPitayaCluster, value)

As with single RedPitaya, but applied to only the master.

RedPitayaDAQServer.SampleChunk — Type

SampleChunk

Struct containing a matrix of samples and associated PerformanceData

Fields

samples::Matrix{Int16}: nxm matrix containing m samples for n channel

performance::Vector{PerformanceData}: PerformanceData object for each RedPitaya that transmitted samples

Missing docstring.

Missing docstring for RedPitayaDAQServer.PerformanceData. Check Documenter's build log for details.

RedPitayaDAQServer.readPipelinedSamples — Function

readPipelinedSamples(rpu::Union{RedPitaya,RedPitayaCluster, RedPitayaClusterView}, wpStart::Int64, numOfRequestedSamples::Int64; chunkSize::Int64 = 25000, rpInfo=nothing)

Request and receive numOfRequestedSamples samples from wpStart on in a pipelined fashion. Return a matrix of samples.

If rpInfo is set to a RPInfo, the PerformanceData sent after every chunkSize samples will be pushed into rpInfo.

readPipelinedSamples(rpu::Union{RedPitaya,RedPitayaCluster, RedPitayaClusterView}, wpStart::Int64, numOfRequestedSamples::Int64, channel::Channel; chunkSize::Int64 = 25000)

Request and receive numOfRequestedSamples samples from wpStart on in a pipelined fashion. The samples and associated PerformanceData are pushed into channel as a SampleChunk.

See SampleChunk.

RedPitayaDAQServer.readFrames — Function

readFrames(rpu::Union{RedPitaya,RedPitayaCluster, RedPitayaClusterView}, startFrame, numFrames, numBlockAverages=1, numPeriodsPerPatch=1; rpInfo=nothing, chunkSize = 50000, useCalibration = false)

Request and receive numFrames frames from startFrame on.

See readPipelinedSamples, convertSamplesToFrames, samplesPerPeriod, periodsPerFrame, updateCalib!.

Arguments

rpu::Union{RedPitaya,RedPitayaCluster, RedPitayaClusterView}: RedPitayas to receive samples from.

startFrame: frame from which to start transmitting

numFrames: number of frames to read

numBlockAverages=1: see convertSamplesToFrames

numPeriodsPerPatch=1: see convertSamplesToFrames

chunkSize=50000: see readPipelinedSamples

rpInfo=nothing: see readPipelinedSamples

useCalibration: convert from Int16 samples to Float32 values based on RedPitayas calibration

Missing docstring.

Missing docstring for RedPitayaDAQServer.convertSamplesToFrames. Check Documenter's build log for details.

RedPitayaDAQServer.calibDACOffset — Function

calibDACOffset(rp::RedPitaya, channel::Integer)

Retrieve the calibration DAC offset for given channel from the RedPitayas EEPROM

RedPitayaDAQServer.calibDACOffset! — Function

calibDACOffset!(rp::RedPitaya, channel::Integer, val)

Store calibration DAC offset val for given channel into the RedPitayas EEPROM. This value is used by the server to offset the output voltage. Absolute value has to be smaller than 1.0 V.

RedPitayaDAQServer.calibADCOffset — Function

calibADCOffset(rp::RedPitaya, channel::Integer)

Retrieve the calibration ADC offset for given channel from the RedPitayas EEPROM.

See also convertSamplesToPeriods,convertSamplesToFrames.

RedPitayaDAQServer.calibADCOffset! — Function

calibADCOffset!(rp::RedPitaya, channel::Integer, val)

Store calibration ADC offset val for given channel into the RedPitayas EEPROM. Absolute value has to be smaller than 1.0 V.

See also convertSamplesToPeriods,convertSamplesToFrames.

RedPitayaDAQServer.calibADCScale — Function

calibADCScale(rp::RedPitaya, channel::Integer)

Retrieve the calibration ADC scale for given channel from the RedPitayas EEPROM.

See also convertSamplesToPeriods,convertSamplesToFrames.

RedPitayaDAQServer.calibADCScale! — Function

calibADCScale(rp::RedPitaya, channel::Integer)

Store calibration ADC scale val for given channel into the RedPitayas EEPROM. Absolute value has to be smaller than 1.0 V.

See also convertSamplesToPeriods,convertSamplesToFrames.